+++ tools/system_libs.py
@@ -112,7 +112,8 @@ def create_lib(libname, inputs):
   suffix = os.path.splitext(libname)[1]
   if suffix in ('.bc', '.o'):
     if len(inputs) == 1:
-      shutil.copyfile(inputs[0], libname)
+      if inputs[0] != libname:
+        shutil.copyfile(inputs[0], libname)
     else:
       shared.Building.link_to_object(inputs, libname)
   elif suffix == '.a':
@@ -311,7 +312,7 @@ class Library(object):
     # libc++-mt.symbols), and if there isn't one, it tries to read the 'default'
     # symbol file, which does not have any optional suffices (e.g.
     # libc++.symbols).
-    basename = os.path.splitext(self.get_filename())[0]
+    basename = shared.unsuffixed(self.get_filename())
     if shared.Settings.WASM_BACKEND:
       symbols_dir = shared.path_from_root('system', 'lib', 'symbols', 'wasm')
     else:
@@ -384,7 +385,7 @@ class Library(object):
     objects = []
     cflags = self.get_cflags()
     for src in self.get_files():
-      o = self.in_temp(os.path.basename(src) + '.o')
+      o = self.in_temp(shared.unsuffixed_basename(src) + '.o')
       commands.append([shared.PYTHON, self.emcc, '-c', src, '-o', o] + cflags)
       objects.append(o)
     run_build_commands(commands)
@@ -1104,7 +1105,7 @@ class libmalloc(MTLibrary, NoBCLibrary):
   def get_default_variation(cls, **kwargs):
     return super(libmalloc, cls).get_default_variation(
       malloc=shared.Settings.MALLOC,
-      is_debug=shared.Settings.DEBUG_LEVEL >= 3,
+      is_debug=shared.Settings.ASSERTIONS >= 2,
       use_errno=shared.Settings.SUPPORT_ERRNO,
       is_tracing=shared.Settings.EMSCRIPTEN_TRACING,
       use_64bit_ops=shared.Settings.MALLOC == 'emmalloc' and (shared.Settings.WASM == 1 or (shared.Settings.WASM_BACKEND and shared.Settings.WASM2JS == 0)),